操作の制限

+++ src/Controller/PostsController.php
@@ -2,6 +2,8 @@
 declare(strict_types = 1);
 namespace App\Controller;
 
+use Cake\Event\EventInterface;
+
 /**
  * Posts Controller
  *
@@ -10,6 +12,23 @@ namespace App\Controller;
  */
 class PostsController extends AppController
 {
+
+    /**
+     * Called before the controller action.
+     * You can use this method to configure and customize components
+     * or perform logic that needs to happen before each controller action.
+     *
+     * @param \Cake\Event\EventInterface $event
+     *            An Event instance
+     * @return \Cake\Http\Response|null|void
+     * @link https://book.cakephp.org/4/en/controllers.html#request-life-cycle-callbacks
+     */
+    public function beforeFilter(EventInterface $event)
+    {
+
+        parent::beforeFilter($event);
+
+    }
 
     /**
      * Index method
@@ -145,6 +164,15 @@ class PostsController extends AppController
             'delete'
         ]);
         $post = $this->Posts->get($id);
+
+        //
+        if ($post->user_id != $this->getIdentityData("id")) {
+            $this->Flash->error(__('Unauthorized Operations.'));
+            return $this->redirect([
+                'action' => 'index'
+            ]);
+        }
+
         if ($this->Posts->delete($post)) {
             $this->Flash->success(__('The post has been deleted.'));
         }
